requirements.txt
Django>=3.1.2
psycopg2>=2.8.6,<2.8.7
django-environ==0.4.5
line-bot-sdk==1.19.0